moved media component files
