+++ src/components/Image.ts
@@ -1,6 +1,6 @@
 import { v4 as uuidv4 } from 'uuid';
-import { Core, MediaComponent } from "./Core";
-import { PostItem } from '../types/post';
+import { Core} from "./Core";
+import { MediaComponent, PostItem } from "../types/post";
 import { IMAGE_INNERHTML } from '../constants/innerHTML';
 
 export class Image extends Core<HTMLElement | HTMLImageElement> implements MediaComponent {
