+++ src/types/post.d.ts
@@ -1,5 +1,3 @@
-import { youtubeRegExp } from "../constants/regExp";
-
 export type PostType = "post" | "image" | "video";
 export type ComponentType = 'media' | 'description' | 'header' | 'footer' | 'modal' | 'form' | 'input';
 export type ModalType = 'view-post-detail' | 'add-card';
@@ -11,6 +9,10 @@ export interface PostItem {
   title: string;
   url?: string;
   description?: string;
+
+  getId(url: string): string;
 }
 
-
+export interface MediaComponent {
+  getId(url?: string): string;
+}
